* release/1.1.0: (28 commits)
  Adding (crude) changelog information for 1.1.0.
  Bumping the version number.
  Removing Pillow from requirements.txt and adding a note about PIL/Pillow to the README. References #72.
  Being a bit friendlier to Python 2.5 users... or is it 2.4. Fixes #82.
  Change handling of spec deletion
  No need to check for None
  Catch KeyError on image.info dictionary
  Using difference instead of subtract
  Indentation tweak
  Makes evaluation of `source_file` lazy
  Test illustrating #75, #74
  Separate create_photo method
  avoid Transpose to crash when exif data doesn't exists
  Allows `None` value for `processors` argument
  Remove _Resize class
  Adds crop module to docs
  Explicitly import crop module
  Renames processor to `TrimBorderColor`
  Gives precedence to user options
  Replaces `quality` argument with `options` dict
  ...

+++ imagekit/models.py
@@ -15,15 +15,15 @@ from imagekit.processors import ProcessorPipeline, AutoConvert
 
 
 class _ImageSpecMixin(object):
-    def __init__(self, processors=None, quality=70, format=None,
+    def __init__(self, processors=None, format=None, options={},
             autoconvert=True):
         self.processors = processors
-        self.quality = quality
         self.format = format
+        self.options = options
         self.autoconvert = autoconvert
 
     def process(self, image, file):
-        processors = ProcessorPipeline(self.processors)
+        processors = ProcessorPipeline(self.processors or [])
         return processors.process(image.copy())
 
 
@@ -35,16 +35,19 @@ class ImageSpec(_ImageSpecMixin):
     """
     _upload_to_attr = 'cache_to'
 
-    def __init__(self, processors=None, quality=70, format=None,
+    def __init__(self, processors=None, format=None, options={},
         image_field=None, pre_cache=False, storage=None, cache_to=None,
         autoconvert=True):
         """
         :param processors: A list of processors to run on the original image.
-        :param quality: The quality of the output image. This option is only
-            used for the JPEG format.
         :param format: The format of the output file. If not provided,
             ImageSpec will try to guess the appropriate format based on the
             extension of the filename and the format of the input image.
+        :param options: A dictionary that will be passed to PIL's
+            ``Image.save()`` method as keyword arguments. Valid options vary
+            between formats, but some examples include ``quality``,
+            ``optimize``, and ``progressive`` for JPEGs. See the PIL
+            documentation for others.
         :param image_field: The name of the model property that contains the
             original image.
         :param pre_cache: A boolean that specifies whether the image should
@@ -71,8 +74,8 @@ class ImageSpec(_ImageSpecMixin):
 
         """
 
-        _ImageSpecMixin.__init__(self, processors, quality=quality,
-                format=format, autoconvert=autoconvert)
+        _ImageSpecMixin.__init__(self, processors, format=format,
+                options=options, autoconvert=autoconvert)
         self.image_field = image_field
         self.pre_cache = pre_cache
         self.storage = storage
@@ -89,12 +92,10 @@ class ImageSpec(_ImageSpecMixin):
 
         # Connect to the signals only once for this class.
         uid = '%s.%s' % (cls.__module__, cls.__name__)
-        post_save.connect(_post_save_handler,
-            sender=cls,
-            dispatch_uid='%s_save' % uid)
-        post_delete.connect(_post_delete_handler,
-            sender=cls,
-            dispatch_uid='%s.delete' % uid)
+        post_save.connect(_post_save_handler, sender=cls,
+                dispatch_uid='%s_save' % uid)
+        post_delete.connect(_post_delete_handler, sender=cls,
+                dispatch_uid='%s.delete' % uid)
 
 
 def _get_suggested_extension(name, format):
@@ -125,6 +126,7 @@ class _ImageSpecFileMixin(object):
         img = open_image(content)
         original_format = img.format
         img = self.field.process(img, self)
+        options = dict(self.field.options or {})
 
         # Determine the format.
         format = self.field.format
@@ -139,29 +141,45 @@ class _ImageSpecFileMixin(object):
                     pass
         format = format or img.format or original_format or 'JPEG'
 
-        if format == 'JPEG':
-            img_to_fobj_kwargs = dict(quality=int(self.field.quality),
-                optimize=True)
-        else:
-            img_to_fobj_kwargs = {}
-
         # Run the AutoConvert processor
         if getattr(self.field, 'autoconvert', True):
             autoconvert_processor = AutoConvert(format)
             img = autoconvert_processor.process(img)
-            img_to_fobj_kwargs.update(autoconvert_processor.save_kwargs)
+            options = dict(autoconvert_processor.save_kwargs.items() + \
+                    options.items())
 
-        imgfile = img_to_fobj(img, format, **img_to_fobj_kwargs)
+        imgfile = img_to_fobj(img, format, **options)
         content = ContentFile(imgfile.read())
         return img, content
 
 
 class ImageSpecFile(_ImageSpecFileMixin, ImageFieldFile):
-    def __init__(self, instance, field, attname, source_file):
+    def __init__(self, instance, field, attname):
         ImageFieldFile.__init__(self, instance, field, None)
-        self.storage = field.storage or source_file.storage
         self.attname = attname
-        self.source_file = source_file
+        self.storage = self.field.storage or self.source_file.storage
+
+    @property
+    def source_file(self):
+        field_name = getattr(self.field, 'image_field', None)
+        if field_name:
+            field_file = getattr(self.instance, field_name)
+        else:
+            image_fields = [getattr(self.instance, f.attname) for f in \
+                    self.instance.__class__._meta.fields if \
+                    isinstance(f, models.ImageField)]
+            if len(image_fields) == 0:
+                raise Exception('{0} does not define any ImageFields, so your '
+                        '{1} ImageSpec has no image to act on.'.format(
+                        self.instance.__class__.__name__, self.attname))
+            elif len(image_fields) > 1:
+                raise Exception('{0} defines multiple ImageFields, but you have '
+                        'not specified an image_field for your {1} '
+                        'ImageSpec.'.format(self.instance.__class__.__name__,
+                        self.attname))
+            else:
+                field_file = image_fields[0]
+        return field_file
 
     def _require_file(self):
         if not self.source_file:
@@ -220,10 +238,11 @@ class ImageSpecFile(_ImageSpecFileMixin, ImageFieldFile):
             self.close()
             del self.file
 
-        try:
-            self.storage.delete(self.name)
-        except (NotImplementedError, IOError):
-            pass
+        if self.name and self.storage.exists(self.name):
+            try:
+                self.storage.delete(self.name)
+            except NotImplementedError:
+                pass
 
         # Delete the filesize cache.
         if hasattr(self, '_size'):
@@ -291,33 +310,11 @@ class _ImageSpecDescriptor(object):
         self.attname = attname
         self.field = field
 
-    def _get_image_field_file(self, instance):
-        field_name = getattr(self.field, 'image_field', None)
-        if field_name:
-            field = getattr(instance, field_name)
-        else:
-            image_fields = [getattr(instance, f.attname) for f in \
-                    instance.__class__._meta.fields if \
-                    isinstance(f, models.ImageField)]
-            if len(image_fields) == 0:
-                raise Exception('{0} does not define any ImageFields, so your '
-                        '{1} ImageSpec has no image to act on.'.format(
-                        instance.__class__.__name__, self.attname))
-            elif len(image_fields) > 1:
-                raise Exception('{0} defines multiple ImageFields, but you have '
-                        'not specified an image_field for your {1} '
-                        'ImageSpec.'.format(instance.__class__.__name__,
-                        self.attname))
-            else:
-                field = image_fields[0]
-        return field
-
     def __get__(self, instance, owner):
         if instance is None:
             return self.field
         else:
-            img_spec_file = ImageSpecFile(instance, self.field,
-                self.attname, self._get_image_field_file(instance))
+            img_spec_file = ImageSpecFile(instance, self.field, self.attname)
             setattr(instance, self.attname, img_spec_file)
             return img_spec_file
 
@@ -358,18 +355,22 @@ class ProcessedImageField(models.ImageField, _ImageSpecMixin):
     _upload_to_attr = 'upload_to'
     attr_class = ProcessedImageFieldFile
 
-    def __init__(self, processors=None, quality=70, format=None,
+    def __init__(self, processors=None, format=None, options={},
         verbose_name=None, name=None, width_field=None, height_field=None,
         autoconvert=True, **kwargs):
         """
         The ProcessedImageField constructor accepts all of the arguments that
         the :class:`django.db.models.ImageField` constructor accepts, as well
-        as the ``processors``, ``format``, and ``quality`` arguments of
+        as the ``processors``, ``format``, and ``options`` arguments of
         :class:`imagekit.models.ImageSpec`.
 
         """
-        _ImageSpecMixin.__init__(self, processors, quality=quality,
-                format=format, autoconvert=autoconvert)
+        if 'quality' in kwargs:
+            raise Exception('The "quality" keyword argument has been'
+                    """ deprecated. Use `options={'quality': %s}` instead.""" \
+                    % kwargs['quality'])
+        _ImageSpecMixin.__init__(self, processors, format=format,
+                options=options, autoconvert=autoconvert)
         models.ImageField.__init__(self, verbose_name, name, width_field,
                 height_field, **kwargs)
